set up basic express server; update dependencies
