--- src/App.jsx
@@ -1,13 +1,0 @@
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        HELLO WORLD!!
-      </header>
-    </div>
-  );
-}
-
-export default App;
